Dobavil builder, mapper, mapstruct, serialazer i mnogo mnogo vsego ostal`nogo ochen` ochen` nugnogo.

+++ src/main/java/ru/skillbox/zerone/backend/repository/CommentRepository.java
@@ -5,14 +5,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skillbox.zerone.backend.model.entity.Comment;
-import ru.skillbox.zerone.backend.model.entity.Post;
-
-import java.util.Optional;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 
-  Page<Comment> findCommentsByPostIdAndParentIsNull(int id, Pageable pageable);
+  Page<Comment> findCommentsByPostIdAndParentIsNull(long id, Pageable pageable);
 
 
 }
